Consolidate Members and Members_Membership (#3161)
I think I refactored this with the expectation that it would be used
somewhere else and it didn’t get there and now it’s just confusing.
Putting it back.
