+++ packages/sdk/src/index.ts
@@ -43,7 +43,6 @@ export * from './streamStateView_GDMChannel'
 export * from './streamStateView_Media'
 export * from './streamStateView_MemberMetadata'
 export * from './streamStateView_Members'
-export * from './streamStateView_Members_Membership'
 export * from './streamStateView_Members_Solicitations'
 export * from './streamStateView_Space'
 export * from './streamStateView_UnknownContent'
